+++ config/webpack.prod.config.js
@@ -1,17 +1,14 @@
-/* eslint-disable import/no-extraneous-dependencies */
-const { join } = require('path')
-
-const { FBS_CONF } = require('./constants')
+const { HOSTING_DIR, CLIENT_ENTRY } = require('./constants')
 const babelConfig = require('./babel.conf.webpackProd.js')
 
 module.exports = {
   mode: 'production',
   entry: [
-    './src/client/index.js',
+    CLIENT_ENTRY,
   ],
   output: {
     filename: 'bundle.js',
-    path: join(process.cwd(), FBS_CONF.hosting.public),
+    path: HOSTING_DIR,
   },
   module: {
     rules: [