+++ helpers/buildBundle.js
@@ -1,0 +1,7 @@
+const webpack = require('webpack')
+
+const config = require('../config/webpack.prod.config.js')
+
+const compiler = webpack(config)
+
+module.exports = () => compiler.run()